Move Merkle verification stuff into new package
