+++ merkle/verifier/hash_chainer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package merkle
+package verifier
 
 import (
 	"github.com/google/trillian/merkle/hashers"
