+++ merkle/verifier/map_verifier_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package merkle
+package verifier
 
 import (
 	"testing"
